refactor of diy phase 1
