+++ src/alpine/CMakeLists.txt
@@ -208,11 +208,14 @@ install(TARGETS alpine
 ################################################
 if(MPI_FOUND)
     set(alpine_par_sources utils/alpine_icet_compositor.cpp
-                           utils/alpine_diy_compositor.cpp)
+                           utils/alpine_diy_compositor.cpp
+                           utils/diy_compositing/alpine_diy_radix_k.cpp)
                          
     set(alpine_par_headers utils/alpine_icet_compositor.hpp
                            utils/alpine_diy_compositor.hpp
-                           utils/alpine_diy_image.hpp)
+                           utils/diy_compositing/alpine_diy_image.hpp
+                           utils/diy_compositing/alpine_diy_collect.hpp
+                           utils/diy_compositing/alpine_diy_radix_k.hpp)
